+++ CHANGELOG.md
@@ -4,7 +4,16 @@
 **New Features:**
 * **[edison-mongo]** Add @ConditionalOnMissingBean to MongoTogglzRepository
 * **[edison-mongo]** Make ID and ETAG constants public to be able to access them from outside
-* **[edison-core]** Allow LDAP authentication for user-defined paths
+* **[edison-core]** Allow LDAP authentication for user-defined paths.
+* **[edison-cache]** It is now possible to configure `CaffeineCache` instances as Spring beans and
+use them using `@Cacheable`. For example, this way it is possible, to configure loading caches.
+The new `example-cache` contains a showcase for this. This feature makes the 'old' `CacheRegistry`
+interface obsolete.
+
+**Deprecations:**
+* **[edison-cache]** The `CacheRegistry` is now deprecated and will be removed in release 2.0.0.
+* **[edison-mongo]** Deprecation of `edison.mongo.passwd`. Use `edison.mongo.password` instead to sanitize value in environment. 
+The support for `edison.mongo.passwd` will be removed in 2.0.0
 
 ## 1.0.0.RC9